Huge refactor, with the implementation of quantityunit as enum. fixed bug where recipe didnt subtract from storage. added the normalize unit logic for both storage, recipe and shoppinglist so kg and g is converted and added or subtraction accordingly etc.

+++ backend/PantioClassLibrary/DTO/ShoppingListDto.cs
@@ -1,5 +1,5 @@
 namespace PantioClassLibrary.DTO;
 
-public record ShoppingListItemDto(Guid Id, string Name, float? Quantity, string? MeasuringUnit, bool IsChecked);
+public record ShoppingListItemDto(Guid Id, string Name, decimal? Quantity, string? MeasuringUnit, bool IsChecked);
 
 public record ShoppingListDto(Guid Id, Guid UserId, string Name, DateTime CreatedAt, List<ShoppingListItemDto> Items);
